+++ LICENSE.md
@@ -1,6 +1,6 @@
 The DynamicTimeWarp.jl package is licensed under the MIT "Expat" License:
 
-> Copyright (c) 2014: Joseph Fowler.
+> Copyright (c) 2016: Alex Williams.
 >
 > Permission is hereby granted, free of charge, to any person obtaining
 > a copy of this software and associated documentation files (the
